Data Annotations e Data Annotations EF

+++ Back/src/ProEventos.API/Controllers/EventoController.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using ProEventos.Domain;
 using ProEventos.Application.Contratos;
 using System;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -26,7 +26,7 @@ namespace ProEventos.API.Controllers
                 var eventos = await _eventoService.GetAllEventosAsync(true);
 
                 if (eventos == null)
-                    return NotFound("Nenhum evento encontrado!");
+                    return NoContent();
 
                 return Ok(eventos);
             }
@@ -45,7 +45,7 @@ namespace ProEventos.API.Controllers
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
 
                 if (evento == null)
-                    return NotFound("Nenhum evento encontrado!");
+                    return NoContent();
 
                 return Ok(evento);
             }
@@ -64,7 +64,7 @@ namespace ProEventos.API.Controllers
                 var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
 
                 if (eventos == null)
-                    return NotFound("Nenhum evento encontrado com o tema informado!");
+                    return NoContent();
 
                 return Ok(eventos);
             }
@@ -76,14 +76,14 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
 
                 if (evento == null)
-                    return BadRequest("Erro ao tentar adicionar o evento!");
+                    return NoContent();
 
                 return Ok(evento);
             }
@@ -95,14 +95,14 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
 
                 if (evento == null)
-                    return BadRequest("Erro ao tentar atualizar o evento!");
+                    return NoContent();
 
                 return Ok(evento);
             }
@@ -118,11 +118,15 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+
+                if (evento == null)
+                    return NoContent();
 
                 if (await _eventoService.DeleteEvento(id))
                     return Ok("Deletado");
                 else
-                    return BadRequest("Erro ao tentar deletar o evento!");
+                    throw new Exception();
             }
             catch (Exception ex)
             {
